refactor: renames, docs
