+++ packages/pure-parse/src/guards/literal.ts
@@ -5,30 +5,30 @@ import { Guard } from './types'
  * Literals types represent single values of primitive types; for example, `true`, `false`, `42`, `"hello"`, and `null` are all types _and_ values.
  * @example
  * ```ts
- * const parseRed = literal('red')
- * const parseOne = literal(1)
+ * const parseRed = literalGuard('red')
+ * const parseOne = literalGuard(1)
  * ```
  * @example
  * Commonly used in discriminated unions:
  * ```ts
- * const isResult = union([
- *  object({
- *    tag: literal('success')
+ * const isResult = unionGuard([
+ *  objectGuard({
+ *    tag: literalGuard('success')
  *  }),
- *  object({
- *    tag: literal('error')
+ *  objectGuard({
+ *    tag: literalGuard('error')
  *   }),
  * ])
  * ```
  * @example
  * If you pass in multiple values, the guard will validate a union type:
  * ```ts
- * const isLogLevel = literal('debug', 'info', 'warning', 'error')
+ * const isLogLevel = literalGuard('debug', 'info', 'warning', 'error')
  * ```
  * @example
  * When explicitly annotating unions, provide a tuple of the union members as type argument:
  * ```ts
- * const isColor = literal<['red', 'green', 'blue']>('red', 'green', 'blue')
+ * const isColor = literalGuard<['red', 'green', 'blue']>('red', 'green', 'blue')
  * ```
  * @param constants compared against `data` with the `===` operator.
  */
