+++ packages/pure-parse/src/guards/union.ts
@@ -1,16 +1,16 @@
 import { Guard } from './types'
 
 /**
- *
+ * Executed `guards` in order and returns true if any guard matches. The result type is a union.
  * @example
  * Commonly used in discriminated unions:
  * ```ts
- * const isResult = union([
- *  object({
- *    tag: literal('success')
+ * const isResult = unionGuard([
+ *  objectGuard({
+ *    tag: literalGuard('success')
  *  }),
- *  object({
- *    tag: literal('error')
+ *  objectGuard({
+ *    tag: literalGuard('error')
  *   }),
  * ])
  * ```
@@ -19,7 +19,7 @@ import { Guard } from './types'
  * ```ts
  * const isId = union<[string, number]>(isString, isNumber)
  * ```
- * Due to a limitation of TypeScript, it is not possible to write `union<string | number>()` or `literal<'red' | 'green' | 'blue'>()`. Therefore, it is generally recommended to omit the type arguments for union types and let TypeScript infer them.
+ * Due to a limitation of TypeScript, it is not possible to write `unionGuard<string | number>()` or `literalGuard<'red' | 'green' | 'blue'>()`. Therefore, it is generally recommended to omit the type arguments for union types and let TypeScript infer them.
  * @param guards any of these guard functions must match the data.
  * @return a guard function that validates unions
  */
