+++ packages/pure-parse/src/parsers/arrays.test.ts
@@ -2,9 +2,9 @@ import { describe, expect, it, test } from 'vitest'
 import { array } from './arrays'
 import { failure, success } from './types'
 
-import { parseNumber, parseString } from './primitives'
-import { withDefault } from './fallback'
 import { literal } from './literal'
+import { oneOf } from './oneOf'
+import { always } from './always'
 
 describe('arrays', () => {
   it('validates when all elements pass validation', () => {
@@ -19,7 +19,7 @@ describe('arrays', () => {
     expect(parseArr(['a', 'b'])).toHaveProperty('tag', 'failure')
   })
   test('with fallbackValue', () => {
-    const parseArr = array(withDefault(literal('#FF0000'), '#00FF00'))
+    const parseArr = array(oneOf(literal('#FF0000'), always('#00FF00')))
     expect(parseArr(['#FF0000', '#FF0000'])).toEqual(
       expect.objectContaining({
         value: ['#FF0000', '#FF0000'],
@@ -40,12 +40,6 @@ describe('arrays', () => {
         value: ['#FF0000', '#00FF00', '#FF0000', '#00FF00'],
       }),
     )
-  })
-  test('that the result type is infallible', () => {
-    const res = withDefault(parseString, '')(123)
-    const a1: typeof res = success('')
-    // @ts-expect-error -- fallbackValue result is infallible
-    const a3: typeof res = failure('')
   })
   describe.todo('self-referential arrays')
 })
