+++ packages/pure-parse/src/parsers/index.ts
@@ -1,8 +1,8 @@
 export * from './types'
 export * from './arrays'
 export * from './object'
-export * from './union'
+export * from './oneOf'
 export * from './primitives'
-export * from './fallback'
+export * from './always'
 export * from './literal'
 export * from './optional'
