+++ packages/pure-parse/src/parsers/object.test.ts
@@ -2,17 +2,17 @@ import { describe, expect, it, test } from 'vitest'
 import { object as objectParseEval, objectNoJit } from './object'
 import { isSuccess, Parser } from './types'
 import type { Equals } from '../internals'
-import { union } from './union'
+import { oneOf } from './oneOf'
 import {
   parseBoolean,
   parseNumber,
   parseString,
   parseUndefined,
 } from './primitives'
-import { withDefault } from './fallback'
 import { Infer } from '../common'
 import { literal } from './literal'
 import { nullable, optional } from './optional'
+import { always } from './always'
 
 const suits = [
   {
@@ -144,7 +144,7 @@ suits.forEach(({ name: suiteName, fn: object }) => {
             'success',
           )
           const parseUnionObj = object({
-            a: union(parseString, parseUndefined),
+            a: oneOf(parseString, parseUndefined),
           })
           expect(parseUnionObj({})).toHaveProperty('tag', 'failure')
           expect(parseUnionObj({ a: undefined })).toHaveProperty(
@@ -237,7 +237,7 @@ suits.forEach(({ name: suiteName, fn: object }) => {
           const parseUser = object({
             id: parseNumber,
             name: parseString,
-            email: optional(withDefault(parseString, defaultEmail)),
+            email: optional(oneOf(parseString, always(defaultEmail))),
           })
           // The email can be a omitted -> Success
           expect(parseUser({ id: 1, name: 'Alice' })).toEqual(
@@ -283,7 +283,7 @@ suits.forEach(({ name: suiteName, fn: object }) => {
           const parseUser = object({
             id: parseNumber,
             name: parseString,
-            email: withDefault(parseString, defaultEmail),
+            email: oneOf(parseString, always(defaultEmail)),
           })
 
           // The property can be a string -> Success
