+++ packages/pure-parse/src/parsers/optional.test.ts
@@ -1,11 +1,12 @@
 import { describe, expect, it } from 'vitest'
 import { nullable, optional, optionalNullable, undefineable } from './optional'
-import { withDefault } from './fallback'
 import { parseString } from './primitives'
+import { oneOf } from './oneOf'
+import { always } from './always'
 
 describe('optional', () => {
   it('works with fallbacks', () => {
-    const parseName = optional(withDefault(parseString, 'anonymous'))
+    const parseName = optional(oneOf(parseString, always('anonymous')))
     expect(parseName('Johannes')).toEqual(
       expect.objectContaining({
         value: 'Johannes',
@@ -25,7 +26,7 @@ describe('optional', () => {
 })
 describe('nullable', () => {
   it('works with fallbacks', () => {
-    const parseName = nullable(withDefault(parseString, 'anonymous'))
+    const parseName = nullable(oneOf(parseString, always('anonymous')))
     expect(parseName('Johannes')).toEqual(
       expect.objectContaining({
         value: 'Johannes',
@@ -45,7 +46,7 @@ describe('nullable', () => {
 })
 describe('undefinable', () => {
   it('works with fallbacks', () => {
-    const parseName = undefineable(withDefault(parseString, 'anonymous'))
+    const parseName = undefineable(oneOf(parseString, always('anonymous')))
     expect(parseName('Johannes')).toEqual(
       expect.objectContaining({
         value: 'Johannes',
@@ -65,7 +66,7 @@ describe('undefinable', () => {
 })
 describe('optionalNullable', () => {
   it('works with fallbacks', () => {
-    const parseName = optionalNullable(withDefault(parseString, 'anonymous'))
+    const parseName = optionalNullable(oneOf(parseString, always('anonymous')))
     expect(parseName('Johannes')).toEqual(
       expect.objectContaining({
         value: 'Johannes',
