+++ packages/pure-parse/src/parsers/optional.ts
@@ -1,7 +1,7 @@
 import { OptionalParser, Parser } from './types'
 import { parseNull, parseUndefined } from './primitives'
 import { optionalSymbol } from '../internals'
-import { union } from './union'
+import { oneOf } from './oneOf'
 
 /**
  * Represent an optional property in an object. It is supposed to be used in combination with `object`.
@@ -24,18 +24,18 @@ import { union } from './union'
  * ```
  * If `email` instead was defined as a union of `string` and `undefined`, the first call to `parseUser` would fail.
  * @param parser A parser to parse the property with.
- * @return a special parser that represents an optional value. If invoked directly, it behaves the same as `union(parseUndefined, parser)`. If invoked by `object`, `object` will treat the property as optional.
+ * @return a special parser that represents an optional value. If invoked directly, it behaves the same as `oneOf(parseUndefined, parser)`. If invoked by `object`, `object` will treat the property as optional.
  */
 export const optional = <T>(parser: Parser<T>): OptionalParser<T> =>
   /*
    * { [optionalValue]: true } is used at runtime by `object` to check if a parser represents an optional value.
    */
-  Object.assign(union(parseUndefined, parser), {
+  Object.assign(oneOf(parseUndefined, parser), {
     [optionalSymbol]: true,
   }) as unknown as OptionalParser<T>
 
 /**
- * Represents a value that can be `null`. Shorthand for `union(parseNull, parser)`.
+ * Represents a value that can be `null`. Shorthand for `oneOf(parseNull, parser)`.
  * @example
  * const parseNullableString = nullable(parseString)
  * parseNullableString(null) // => ParseSuccess<string | null>
@@ -43,10 +43,10 @@ export const optional = <T>(parser: Parser<T>): OptionalParser<T> =>
  * @param parser a parser function.
  */
 export const nullable = <T>(parser: Parser<T>): Parser<T | null> =>
-  union(parseNull, parser)
+  oneOf(parseNull, parser)
 
 /**
- * Represents a value that can be `undefined`. Shorthand for `union(parseUndefined, parser)`.
+ * Represents a value that can be `undefined`. Shorthand for `oneOf(parseUndefined, parser)`.
  * @example
  * const parseUndefineableString = undefineable(parseString)
  * parseUndefineableString(undefined) // => ParseSuccess<string | undefined>
@@ -54,10 +54,10 @@ export const nullable = <T>(parser: Parser<T>): Parser<T | null> =>
  * @param parser
  */
 export const undefineable = <T>(parser: Parser<T>): Parser<T | undefined> =>
-  union(parseUndefined, parser)
+  oneOf(parseUndefined, parser)
 
 /**
- * Represents an optional property that can also be null. Shorthand for `optional(union(parseNull, parser))`.
+ * Represents an optional property that can also be null. Shorthand for `optional(oneOf(parseNull, parser))`.
  * @param parser a parser function.
  */
 export const optionalNullable = <T>(
